requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg[binary]==3.1.13
pydantic==2.5.0
python-multipart==0.0.6
pyyaml==6.0.1
openapi-spec-validator==0.7.1
httpx==0.25.2
pytest==7.4.3
openai==1.3.7
anthropic==0.7.8
